+++ src/events/guildMemberAdd.js
@@ -3,8 +3,8 @@ const { startVerification } = require('../services/verification');
 
 module.exports = {
   name: Events.GuildMemberAdd,
-  async execute(member) {
+  async execute(member, client) {
     console.log(`New member joined: ${member.user.tag} in ${member.guild.name}`);
-    await startVerification(member);
+    await startVerification(member, client);
   },
 };